Default OTA off, add switch to turn on and UI improvements
At startup, connects to WiFi to get NTP time only and then disconnects and turns off WiFi. This lessens current consumption of Tiger. A UI indication has been added to show when either of the reed switches are being pressed with a timer of how many seconds it has been pressed for. If second button is pressed for 5 seconds on clock screen then makes attempt to connect to WiFi and enable OTA.